Factor out Semigroup classes into separate internal module
This will be needed to avoid import cycles with `Data.Ord.Singletons` in a
subsequent commit.
