+++ singletons-base/src/Data/List/Singletons/Internal.hs
@@ -31,7 +31,8 @@ import Data.Eq.Singletons
 import Data.Maybe
 import Data.Maybe.Singletons
 import Data.Ord.Singletons
-import Data.Semigroup.Singletons.Internal.Wrappers (SSemigroup(..), type (<>@#@$))
+import Data.Semigroup.Singletons.Internal.Classes (SSemigroup(..), type (<>@#@$))
+import Data.Semigroup.Singletons.Internal.Wrappers ()
 import Data.Singletons.Base.Instances
 import Data.Singletons.TH
 import Data.Tuple.Singletons
