+++ singletons-base/src/Data/Semigroup/Singletons.hs
@@ -64,6 +64,7 @@ import Data.Ord.Singletons hiding
 import Data.Ord.Singletons.Disambiguation
 import qualified Data.Semigroup as Semi (Min(..), Max(..))
 import Data.Semigroup (First(..), Last(..), WrappedMonoid(..), Arg(..))
+import Data.Semigroup.Singletons.Internal.Classes
 import Data.Semigroup.Singletons.Internal.Wrappers
 import Data.Singletons.Base.Enum
 import Data.Singletons.Base.Instances
